class_room_backend/requirements.txt
aioredis==1.3.1
amqp==5.0.2
asgiref==3.3.0
async-timeout==3.0.1
attrs==20.3.0
autobahn==20.7.1
Automat==20.2.0
billiard==3.6.3.0
blis==0.7.4
catalogue==1.0.0
celery==5.0.5
certifi==2020.12.5
cffi==1.14.4
channels==3.0.2
channels-redis==3.2.0
chardet==4.0.0
ChatterBot==1.0.8
chatterbot-corpus==1.2.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
constantly==15.1.0
cryptography==3.2.1
cymem==2.0.5
daphne==3.0.1
dj-database-url==0.5.0
Django==3.1.3
django-celery-results==2.0.0
django-cors-headers==3.5.0
django-filter==2.4.0
django-heroku==0.3.1
django-multipleselectfield==0.2.0.post2
djangorestframework==3.12.1
djangorestframework-bulk==0.2.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz
gunicorn==20.0.4
hiredis==1.1.0
hyperlink==20.0.1
idna==2.10
importlib-metadata==3.3.0
incremental==17.5.0
joblib==1.0.0
kombu==5.0.2
Markdown==3.3.3
mathparse==0.1.2
msgpack==1.0.0
murmurhash==1.0.5
nltk==3.5
numpy==1.19.4
plac==1.1.3
preshed==3.0.5
prompt-toolkit==3.0.8
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==20.0.0
python-dateutil==2.8.1
pytz==2020.4
PyYAML==3.13
redis==3.5.3
regex==2020.11.13
requests==2.25.1
service-identity==18.1.0
six==1.15.0
spacy==2.3.5
SQLAlchemy==1.3.22
sqlparse==0.4.1
srsly==1.0.5
thinc==7.4.5
tqdm==4.54.1
Twisted==20.3.0
txaio==20.4.1
typing-extensions==3.7.4.3
urllib3==1.26.2
vine==5.0.0
wasabi==0.8.0
wcwidth==0.2.5
whitenoise==5.2.0
zipp==3.4.0
zope.interface==5.2.0

